api-gateway/requirements.txt
flask
requests
python-dotenv
flask-CORS
PyJWT
Flask-SocketIO
python-socketio
websocket-client